Simplified and moved around some more planner code
