+++ tidb/expression/expression.go
@@ -42,16 +42,6 @@ const (
 	scalarFunctionFlag byte = 3
 )
 
-// EvalAstExpr evaluates ast expression directly.
-// Note: initialized in planner/core
-// import expression and planner/core together to use EvalAstExpr
-var EvalAstExpr func(sctx sessionctx.Context, expr ast.ExprNode) (types.Datum, error)
-
-// RewriteAstExpr rewrites ast expression directly.
-// Note: initialized in planner/core
-// import expression and planner/core together to use EvalAstExpr
-var RewriteAstExpr func(sctx sessionctx.Context, expr ast.ExprNode, schema *Schema, names types.NameSlice) (Expression, error)
-
 // VecExpr contains all vectorized evaluation methods.
 type VecExpr interface {
 	// Vectorized returns if this expression supports vectorized evaluation.
