+++ tidb/planner/physical_hashagg.go
@@ -155,3 +155,32 @@ func (p *basePhysicalAgg) getAggFuncCostFactor(isMPP bool) (factor float64) {
 	}
 	return
 }
+
+// ResolveIndices implements Plan interface.
+func (p *basePhysicalAgg) ResolveIndices() (err error) {
+	err = p.physicalSchemaProducer.ResolveIndices()
+	if err != nil {
+		return err
+	}
+	for _, aggFun := range p.AggFuncs {
+		for i, arg := range aggFun.Args {
+			aggFun.Args[i], err = arg.ResolveIndices(p.children[0].Schema())
+			if err != nil {
+				return err
+			}
+		}
+		for _, byItem := range aggFun.OrderByItems {
+			byItem.Expr, err = byItem.Expr.ResolveIndices(p.children[0].Schema())
+			if err != nil {
+				return err
+			}
+		}
+	}
+	for i, item := range p.GroupByItems {
+		p.GroupByItems[i], err = item.ResolveIndices(p.children[0].Schema())
+		if err != nil {
+			return err
+		}
+	}
+	return
+}
