+++ tidb/planner/physical_schemaproducer.go
@@ -1,0 +1,50 @@
+//
+// This source code is a modified form of original source from the TiDB project, which has the following copyright header(s):
+//
+
+// Copyright 2017 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package planner
+
+import "github.com/squareup/pranadb/tidb/expression"
+
+// physicalSchemaProducer stores the schema for the physical plans who can produce schema directly.
+type physicalSchemaProducer struct {
+	schema *expression.Schema
+	basePhysicalPlan
+}
+
+// Schema implements the Plan.Schema interface.
+func (s *physicalSchemaProducer) Schema() *expression.Schema {
+	if s.schema == nil {
+		if len(s.Children()) == 1 {
+			// default implementation for plans has only one child: proprgate child schema.
+			// multi-children plans are likely to have particular implementation.
+			s.schema = s.Children()[0].Schema().Clone()
+		} else {
+			s.schema = expression.NewSchema()
+		}
+	}
+	return s.schema
+}
+
+// SetSchema implements the Plan.SetSchema interface.
+func (s *physicalSchemaProducer) SetSchema(schema *expression.Schema) {
+	s.schema = schema
+}
+
+func (p *physicalSchemaProducer) ResolveIndices() (err error) {
+	err = p.basePhysicalPlan.ResolveIndices()
+	return err
+}
