--- tidb/planner/resolve_indices.go
@@ -1,83 +1,0 @@
-//
-// This source code is a modified form of original source from the TiDB project, which has the following copyright header(s):
-//
-
-// Copyright 2016 PingCAP, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package planner
-
-// ResolveIndices implements Plan interface.
-func (p *PhysicalUnionScan) ResolveIndices() (err error) {
-	err = p.basePhysicalPlan.ResolveIndices()
-	if err != nil {
-		return err
-	}
-	for i, expr := range p.Conditions {
-		p.Conditions[i], err = expr.ResolveIndices(p.children[0].Schema())
-		if err != nil {
-			return err
-		}
-	}
-	resolvedHandleCol, err := p.HandleCols.ResolveIndices(p.children[0].Schema())
-	if err != nil {
-		return err
-	}
-	p.HandleCols = resolvedHandleCol
-	return
-}
-
-// ResolveIndices implements Plan interface.
-func (p *basePhysicalAgg) ResolveIndices() (err error) {
-	err = p.physicalSchemaProducer.ResolveIndices()
-	if err != nil {
-		return err
-	}
-	for _, aggFun := range p.AggFuncs {
-		for i, arg := range aggFun.Args {
-			aggFun.Args[i], err = arg.ResolveIndices(p.children[0].Schema())
-			if err != nil {
-				return err
-			}
-		}
-		for _, byItem := range aggFun.OrderByItems {
-			byItem.Expr, err = byItem.Expr.ResolveIndices(p.children[0].Schema())
-			if err != nil {
-				return err
-			}
-		}
-	}
-	for i, item := range p.GroupByItems {
-		p.GroupByItems[i], err = item.ResolveIndices(p.children[0].Schema())
-		if err != nil {
-			return err
-		}
-	}
-	return
-}
-
-func (p *physicalSchemaProducer) ResolveIndices() (err error) {
-	err = p.basePhysicalPlan.ResolveIndices()
-	return err
-}
-
-// ResolveIndices implements Plan interface.
-func (p *basePhysicalPlan) ResolveIndices() (err error) {
-	for _, child := range p.children {
-		err = child.ResolveIndices()
-		if err != nil {
-			return err
-		}
-	}
-	return
-}
